// src/lib/api/models/updatePropertyDtoVerificationStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * RealHub API
 * Multi-tenant Real Estate Platform API
 * OpenAPI spec version: v1
 */

export type UpdatePropertyDtoVerificationStatus = typeof UpdatePropertyDtoVerificationStatus[keyof typeof UpdatePropertyDtoVerificationStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UpdatePropertyDtoVerificationStatus = {
  DRAFT: 'DRAFT',
  PENDING: 'PENDING',
  VERIFIED: 'VERIFIED',
  REJECTED: 'REJECTED',
} as const;